+++ API/src/Modules/XpSystems/Momentum.XpSystems.Application/DTOs/Cosmetic/LevelsDto.cs
@@ -6,10 +6,10 @@
         public int StartingValue { get; set; }
         public int LinearScaleBaseIncrease { get; set; }
         public int LinearScaleInterval { get; set; }
-        public float LinearScaleIntervalMultiplier { get; set; }
+        public double LinearScaleIntervalMultiplier { get; set; }
         public int StaticScaleStart { get; set; }
-        public float StaticScaleBaseMultiplier { get; set; }
+        public double StaticScaleBaseMultiplier { get; set; }
         public int StaticScaleInterval { get; set; }
-        public float StaticScaleIntervalMultiplier { get; set; }
+        public double StaticScaleIntervalMultiplier { get; set; }
     }
 }
