+++ API/src/Modules/XpSystems/Momentum.XpSystems.Core/Models/Rank/Top10.cs
@@ -3,6 +3,6 @@
     public class Top10
     {
         public int WorldRecordPoints { get; set; }
-        public float[] RankPercentages { get; set; }
+        public double[] RankPercentages { get; set; }
     }
 }
